Add Day 2.1 updated dive logic
The new dive logic changes how depth is calculated. We need a new
property, "Aim", to determine the dive rate for each forward movement.
Depth is no longer a straight assignment, but is calculated based on
"aim" and how far forward we have moved. Up/Down now control the value
of "Aim" and don't take effect until we move forward.

The legacy "Move" implementation is renamed to "SimpleMove" for
backwards compatibility, and "Move" becomes the new correct logic for
depth calculation.

A new Reset() method is added as well so we don't have to create a bunch
of independent submarines in Main.