Add PowerDiagnostic record
We need to be able to parse a list of binary numbers, fuck around
counting some bits, and then generate two more binary numbers based on
the most- and least-frequent bit in each position.

This was a nightmare because there appears to be no way to convert a
BitArray into its integer representation without FUCKING TURNING IT INTO
A STRING. I am so fucking tired of this.