Reintroduce a solution and sub-projects
As the codebase grows it's clear that a flat directory full of junk is
getting pretty messy. It's time to start separating our concerns into
their own packages.

This also put the spotlight on a couple areas around the `Ingestor` that
broke a few barriers with things being a bit too oddly-coupled. We'll
fix those up later.