Add basic Bingo board and cell records
This adds the very basic mechanics of Bingo in the form of two records:
boards and cells.

A board represents a player's bingo card, and is the outermost API for a
player in the context of the game. A board has a 2-dimensional List of
cells, which are instances of a number on their bingo card. A Cell has
some minimal internal state to know its number and whether or not it has
been marked.

This allows the Board to not need to have direct knowledge of the
numbers drawn, as we can just have it relay which numbers have been
drawn to each cell, and then blindly check which ones are marked.

This is similar to real-life bingo, where the card is marked by the
player and not the game officiator.